requirements.txt
flask
pyodbc
python-dotenv
jinja2
